move internal/ folder to api/
