+++ packages/serverless-api/src/api/assets.ts
